Finish auth microservice

+++ pkg/models/auth.go
@@ -1,7 +1,7 @@
 package models
 
 type User struct {
-	Id       int    `json:"id" gorm:"primary_key"`
+	Id       int64  `json:"id" gorm:"primary_key"`
 	Username string `json:"username"`
 	Password string `json:"password"`
 }
